requirements.txt
cvxpy
tensorboard
geopy
scikit-learn
networkx == 2.4
numpy
scipy
matplotlib
jupyter
torchtext==0.5.0
spacy==2.0.0
mplleaflet